Refactoring The Builder Class of SliderPager with best practices
