+++ appintro/src/main/java/com/github/paolorotolo/appintro/model/SliderPage.kt
@@ -46,4 +46,105 @@ data class SliderPage @JvmOverloads constructor(
         newBundle.putInt(ARG_BG_DRAWABLE, this.bgDrawable)
         return newBundle
     }
+
+    class Builder {
+
+        private var title: CharSequence? = null
+
+        private var description: CharSequence? = null
+
+        @DrawableRes
+        private var imageDrawable: Int = 0
+
+        @ColorInt
+        private var bgColor: Int = 0
+
+        @DrawableRes
+        private var bgDrawable: Int = 0
+
+        @ColorInt
+        private var titleColor: Int = 0
+
+        @ColorInt
+        private var descColor: Int = 0
+
+        private var titleTypeface: String? = null
+
+        private var descTypeface: String? = null
+
+        @FontRes
+        private var titleTypefaceRes: Int = 0
+
+        @FontRes
+        private var descTypefaceRes: Int = 0
+
+        fun title(title: CharSequence): Builder {
+            this.title = title
+            return this
+        }
+
+        fun description(description: CharSequence): Builder {
+            this.description = description
+            return this
+        }
+
+        fun imageDrawable(imageDrawable: Int): Builder {
+            this.imageDrawable = imageDrawable
+            return this
+        }
+
+        fun bgColor(bgColor: Int): Builder {
+            this.bgColor = bgColor
+            return this
+        }
+
+        fun bgDrawable(bgDrawable: Int): Builder {
+            this.bgDrawable = bgDrawable
+            return this
+        }
+
+        fun titleColor(titleColor: Int): Builder {
+            this.titleColor = titleColor
+            return this
+        }
+
+        fun descColor(descColor: Int): Builder {
+            this.descColor = descColor
+            return this
+        }
+
+        fun titleTypeface(titleTypeface: String): Builder {
+            this.titleTypeface = titleTypeface
+            return this
+        }
+
+        fun titleTypefaceRes(@FontRes titleTypefaceRes: Int): Builder {
+            this.titleTypefaceRes = titleTypefaceRes
+            return this
+        }
+
+        fun descTypeface(descTypeface: String): Builder {
+            this.descTypeface = descTypeface
+            return this
+        }
+
+        fun descTypefaceRes(@FontRes descTypefaceRes: Int): Builder {
+            this.descTypefaceRes = descTypefaceRes
+            return this
+        }
+
+        fun build() = SliderPage(
+                title = this.title,
+                description = this.description,
+                imageDrawable = this.imageDrawable,
+                bgDrawable = this.bgDrawable,
+                bgColor = this.bgColor,
+                titleColor = this.titleColor,
+                descColor = this.descColor,
+                titleTypeface = this.titleTypeface,
+                titleTypefaceFontRes = this.titleTypefaceRes,
+                descTypeface = this.descTypeface,
+                descTypefaceFontRes = this.descTypefaceRes
+        )
+    }
 }
